Functions for rectangular domain predictions (#27)
Add routines for doing predictions, transforms and plots on 2D rectangular domains. Move all luigi Tasks for pipeline of rectpred into separate modules so that we can access the underlying functions without having luigi installed.

Also includes minor tweak to `get_embeddings` util: Add flexibility for datasets created on-the-fly (for example from a
single image) which don't have a data-path.
