+++ convml_tt/utils.py
@@ -29,10 +29,13 @@ def get_embeddings(tile_dataset: ImageSingletDataset, model, prediction_batch_si
 
     dims = ("tile_id", "emb_dim")
     coords = dict(tile_id=tile_ids)
-    attrs = dict(
-        data_dir=str(Path(tile_dataset.data_dir).absolute()),
-        tile_type=tile_dataset.tile_type.name,
-        stage=tile_dataset.stage,
-    )
+
+    attrs = {}
+    if hasattr(tile_dataset, "tile_type"):
+        attrs["tile_type"] = tile_dataset.tile_type.name
+    if hasattr(tile_dataset, "stage"):
+        attrs["stage"] = tile_dataset.stage
+    if hasattr(tile_dataset, "data_dir"):
+        attrs["data_dir"] = str(Path(tile_dataset.data_dir).absolute())
 
     return xr.DataArray(embeddings, dims=dims, coords=coords, attrs=attrs)
