+++ setup.cfg
@@ -34,6 +34,7 @@ install_requires =
     seaborn
     parse
     jupyter
+    scikit-image
 setup_requires =
     setuptools >= 40.4
     setuptools_scm
